+++ include/bmx/apps/AppUtils.h
@@ -115,7 +115,7 @@ bool parse_klv_opt(const char *klv_opt_str, mxfKey *key, uint32_t *track_num);
 bool parse_anc_data_types(const char *types_str, std::set<ANCDataType> *types);
 bool parse_checksum_type(const char *type_str, ChecksumType *type);
 bool parse_rdd6_lines(const char *lines_str, uint16_t *lines);
-bool parse_track_indexes(const char *tracks_str, std::set<uint32_t> *track_indexes);
+bool parse_track_indexes(const char *tracks_str, std::set<size_t> *track_indexes);
 bool parse_mxf_auid(const char *mxf_auid_str, UL *mxf_auid);
 bool parse_bytes_size(const char *size_str, int64_t *size_out);
 bool parse_signal_standard(const char *str, MXFSignalStandard *value);
